Remove the special behavior for the event's address, this just makes things soo more complicated.
Relying on having to use the <address> tag sounds nice at first sight but after all, if we derive and reconstruct the HTML out of it, this has many drawbacks.
Such as 1) to know what your page will look like you have to understand the webcomponent 2) you rely on what DOM nodes the webcomp inserts in order to know how to style them
and 3) it makes the <address> tag required, even though it is not really useful, the better way is schema.org outlines of the content.
Additionally the <hc-event> component triggers a massive relayout on a page with many events, that is not really desirable.